Update Storybook config and add preview setup
