+++ .storybook/main.ts
@@ -2,15 +2,16 @@ import type { StorybookConfig } from "@storybook/nextjs";
 
 const config: StorybookConfig = {
   stories: [
-    "../stories/**/*.mdx",
-    "../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)",
+    "../components/**/*.mdx",
+    "../components/**/*.stories.@(js|jsx|mjs|ts|tsx)",
   ],
   addons: [
-    "@storybook/addon-onboarding",
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@chromatic-com/storybook",
+    "@storybook/addon-onboarding",
     "@storybook/addon-interactions",
+    "@storybook/addon-toolbars",
+    "@chromatic-com/storybook",
   ],
   framework: {
     name: "@storybook/nextjs",
